pyproject.toml
[tool.poetry]
name = "splai"
version = "0.1.0"
description = "An AI helper"
authors = ["Rich Dougherty <rich@rd.nz>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.129"
openai = "^0.27.2"
gitpython = "^3.1.31"
chromadb = "^0.3.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
